+++ src/main/webapp/branch-sales/delete.jsp
@@ -1,18 +1,31 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
-<%@ page import="test.dao.SalesDao" %>
+<%@ page import="test.dao.BranchSalesDao" %>
 <%
-    request.setCharacterEncoding("UTF-8");
+request.setCharacterEncoding("UTF-8");
 
-    String branchId = (String)session.getAttribute("branchId");
-    if(branchId == null){
-        response.sendRedirect(request.getContextPath()+"/userp/branchlogin-form.jsp");
-        return;
-    }
-
+    // 삭제할 매출 번호와 지점 ID 파라미터
     int salesId = Integer.parseInt(request.getParameter("salesId"));
-    boolean isSuccess = SalesDao.getInstance().delete(salesId, branchId);
+    String branchId = request.getParameter("branchId");
+
+    // DB 삭제 수행
+    boolean isSuccess = BranchSalesDao.getInstance().delete(salesId, branchId);
 %>
-<script>
-    alert("<%= isSuccess ? "삭제 성공" : "삭제 실패" %>");
-    location.href = "list.jsp";
-</script>
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<title>/branch-sales/delete.jsp</title>
+</head>
+<body>
+    <script>
+        <%-- 삭제 성공 여부에 따라 알림창 후 이동 처리 --%>
+        <% if(isSuccess){ %>
+            alert("삭제 성공");
+            location.href="<%=request.getContextPath()%>/branch-sales/list.jsp";
+        <% } else { %>
+            alert("삭제 실패");
+            history.back();
+        <% } %>
+    </script>
+</body>
+</html>
