+++ src/main/webapp/branch-sales/detail.jsp
@@ -1,18 +1,35 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
-<%@ page import="test.dao.SalesDao" %>
-<%@ page import="test.dto.SalesDto" %>
+<%@ page import="test.dao.BranchSalesDao" %>
+<%@ page import="test.dto.BranchSalesDto" %>
 <%
-    request.setCharacterEncoding("UTF-8");
+request.setCharacterEncoding("UTF-8");
 
+    // 로그인 세션 확인
     String branchId = (String)session.getAttribute("branchId");
     if(branchId == null){
-        response.sendRedirect(request.getContextPath()+"/userp/branchlogin-form.jsp");
+        response.sendRedirect(request.getContextPath() + "/userp/branchlogin-form.jsp");
         return;
     }
 
+    // 매출 ID 파라미터
     int salesId = Integer.parseInt(request.getParameter("salesId"));
-    SalesDto dto = SalesDao.getInstance().getById(salesId, branchId);
 
+    // 매출 정보 조회
+    BranchSalesDto dto = BranchSalesDao.getInstance().getById(salesId, branchId);
+
+    String cpath = request.getContextPath();
+%>
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<title>/branch-sales/detail.jsp</title>
+<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
+</head>
+<body>
+<div class="container mt-3">
+
+<%
     if(dto == null){
 %>
     <script>
@@ -20,31 +37,31 @@
         history.back();
     </script>
 <%
-        return;
-    }
-
-    String cpath = request.getContextPath();
+    } else {
 %>
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <title>/branch-sales/detail.jsp</title>
-    <script>const cpath = "<%= cpath %>";</script>
-</head>
-<body>
+
     <h2>매출 상세 보기</h2>
-    <table border="1" cellpadding="10">
-        <tr><th>매출 ID</th><td><%= dto.getSalesId() %></td></tr>
-        <tr><th>지점 ID</th><td><%= dto.getBranchId() %></td></tr>
-        <tr><th>매출 발생일</th><td><%= dto.getCreated_at() %></td></tr>
-        <tr><th>총 매출 금액</th><td><%= dto.getTotalAmount() %> 원</td></tr>
+
+    <!-- 매출 상세 테이블 -->
+    <table class="table table-bordered mt-3">
+        <tr><th class="table-light" style="width: 150px;">매출 ID</th><td><%= dto.getSalesId() %></td></tr>
+        <tr><th class="table-light">지점 ID</th><td><%= dto.getBranchId() %></td></tr>
+        <tr><th class="table-light">매출 발생일</th><td><%= dto.getCreated_at() %></td></tr>
+        <tr><th class="table-light">총 매출 금액</th><td><%= dto.getTotalAmount() %> 원</td></tr>
     </table>
 
-    <p style="margin-top: 15px;">
-        <button onclick="location.href = cpath + '/branch-sales/update-form.jsp?salesId=<%= dto.getSalesId() %>'">수정</button>
-        <button onclick="location.href = cpath + '/branch-sales/delete.jsp?salesId=<%= dto.getSalesId() %>'">삭제</button>
-        <button onclick="location.href = cpath + '/branch-sales/list.jsp'">목록</button>
-    </p>
+    <!-- 버튼 영역 -->
+    <div class="mt-3">
+        <a href="<%= cpath %>/branch-sales/update-form.jsp?salesId=<%= dto.getSalesId() %>" class="btn btn-warning">수정</a>
+        <a href="<%= cpath %>/branch-sales/delete.jsp?salesId=<%= dto.getSalesId() %>&branchId=<%= dto.getBranchId() %>" 
+           class="btn btn-danger" onclick="return confirm('정말 삭제하시겠습니까?');">삭제</a>
+        <a href="<%= cpath %>/branch-sales/list.jsp" class="btn btn-secondary">목록</a>
+    </div>
+
+<%
+    }
+%>
+
+</div>
 </body>
 </html>
