+++ src/main/webapp/branch-sales/update-form.jsp
@@ -1,18 +1,33 @@
 <%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
-<%@ page import="test.dao.SalesDao" %>
-<%@ page import="test.dto.SalesDto" %>
+<%@ page import="test.dao.BranchSalesDao" %>
+<%@ page import="test.dto.BranchSalesDto" %>
 <%
-    request.setCharacterEncoding("UTF-8");
+request.setCharacterEncoding("UTF-8");
 
+    // 로그인 세션 확인
     String branchId = (String)session.getAttribute("branchId");
     if(branchId == null){
-        response.sendRedirect(request.getContextPath()+"/userp/branchlogin-form.jsp");
+        response.sendRedirect(request.getContextPath() + "/userp/branchlogin-form.jsp");
         return;
     }
 
+    // 매출 정보 조회
     int salesId = Integer.parseInt(request.getParameter("salesId"));
-    SalesDto dto = SalesDao.getInstance().getById(salesId, branchId);
+    BranchSalesDto dto = BranchSalesDao.getInstance().getById(salesId, branchId);
 
+    String cpath = request.getContextPath();
+%>
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<title>/branch-sales/update-form.jsp</title>
+<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
+</head>
+<body>
+<div class="container mt-3">
+
+<%
     if(dto == null){
 %>
     <script>
@@ -20,28 +35,31 @@
         history.back();
     </script>
 <%
-        return;
-    }
+    } else {
 %>
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <title>/branch-sales/update-form.jsp</title>
-</head>
-<body>
+
     <h2>매출 수정</h2>
-    <form action="update.jsp" method="post">
+
+    <form action="<%=cpath%>/branch-sales/update.jsp" method="post" class="mt-3" style="max-width:400px;">
         <input type="hidden" name="salesId" value="<%= dto.getSalesId() %>">
-        <table border="1" cellpadding="10">
+
+        <table class="table table-bordered">
             <tr>
-                <th>총 매출 금액</th>
-                <td><input type="number" name="totalAmount" value="<%= dto.getTotalAmount() %>" required></td>
+                <th class="table-light" style="width:150px;">총 매출 금액</th>
+                <td><input type="number" name="totalAmount" class="form-control" value="<%= dto.getTotalAmount() %>" required></td>
             </tr>
         </table>
-        <br>
-        <button type="submit">수정</button>
-        <button type="button" onclick="location.href='list.jsp'">취소</button>
+
+        <div class="mt-3">
+            <button type="submit" class="btn btn-warning">수정</button>
+            <a href="<%=cpath%>/branch-sales/list.jsp" class="btn btn-secondary">취소</a>
+        </div>
     </form>
+
+<%
+    }
+%>
+
+</div>
 </body>
 </html>
